NULL value inside UserVisitLog.php causes error
Hey, I have the following problem while doing acceptance tests: The login is not working because the codecept browser language is not set. This will cause an integrity constraint violation because the UserVisitLog.php is setting the language value to NULL but the database does not accept NULL values.

So you should change the UserVisitLog.php as suggested or the user_visit_log should accept NULL values by default migration.

PS: I know you can change the HTTP_ACCEPT_LANGUAGE from the PhpBrowser (I have done it) but this is a potential error source for users. Just wanted to let you know. :)